added some extremely basic ID support
When a stack file is exported and then imported the IDs should now be the same. When a stack file is merged, however, it's extremely likely there will be a conflict on the ID of objects if both branches have created new objects. When this happens lcVCS will not be able to set the ID of the conflicted control correctly and will fail silently. With the current implementation one failure may lead to other objects created after it failing because the stack file is not re-created in ID order and the ID assigned to the conflicted control may be that required for another control which could start a chain of failures.